fix(cron): align retry-config typing with upstream default-knob removal
Upstream removed the configurable cron.retry knob (CronRetryOn export and
CronConfig.retry) in favor of built-in defaults. Adopt the current
resolveRetryConfig() zero-arg shape and import CronRetryOn from retry-hint,
thread the required run origin through the invalid-persisted-manual skip
paths, and split the oversized ops regression suite so it stays under the
max-lines ratchet.
